requirements.txt
google-genai==1.24.0
python-dotenv==1.0.1
pytz==2025.1
schedule==1.2.2
tweepy==4.15.0
slack-sdk==3.23.0
flask==3.0.0